backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
networkx>=3.2
gitpython>=3.1
tree-sitter>=0.21,<0.23
tree-sitter-languages>=1.10
pydantic>=2.5
pydantic-settings>=2.1
python-multipart>=0.0.6
pytest>=8.0
httpx>=0.27
faiss-cpu>=1.7
sentence-transformers>=2.2
anthropic>=0.39
aiosqlite>=0.19
PyJWT>=2.8
cryptography>=42.0
openai>=1.30
google-genai>=1.0
slowapi>=0.1.9
neo4j>=5.0
